fix: restore 0-if-polymorphic fallback for nearly-empty virtual bases
Codex review found a real regression, not a missing enhancement:
struct E : virtual A { virtual void e(); }; where A has no data
members of its own ("nearly empty" for the Itanium ABI's
virtual-primary-base rule) can be laid out so E and A share one vptr
slot at offset 0, with GCC emitting no local _vptr.E member for E at
all -- confirmed by DIE dump, and different from every other
virtual-base case this fix already covered (a class with a real data
member of its own, like the existing struct D : virtual A { virtual
void d(); int di; };, DOES get its own local vptr member).

Because E's only base is virtual, it's entirely excluded from the
primary-base walk (mirroring base_offsets's own long-standing
virtual-base exclusion -- the offset is dynamic, not static), so E
fell through to None where the pre-fix code -- which never
distinguished "resolved via a real mechanism" from "just assumed 0
whenever vtable is non-empty" -- used to give the correct answer 0.

Fixed with a final fallback pass in _finalize_vptr_offsets, after the
fixed-point resolution loop completes: any type still None with a
non-empty vtable is set to 0, restoring the original blanket
heuristic exactly for the residual set this fix's more precise
mechanisms cannot explain. Strictly additive over the pre-fix
behavior: every case the old code resolved to 0 still resolves to at
worst 0; several cases it could only guess at now resolve to a real,
DWARF-derived offset instead.

New regression test with exactly this shape. Splits the growing
TestDwarfSnapshotVptrOffset class (all vptr coverage added across
this fix's several review rounds) out of test_dwarf_snapshot.py into
a new sibling file, tests/test_dwarf_vptr_offset.py -- the fourth
regression test pushed the parent file to 2045 lines, over the
2000-line file-size hard cap (scripts/check_ai_readiness.py).
